backend/pyproject.toml
[tool.poetry]
name = "todo-backend"
version = "0.1.0"
description = "Backend for the Todo application"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
sqlmodel = "^0.0.22"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
passlib = {extras = ["bcrypt"], version = "^2.0.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic-settings = "^2.6.1"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
python-multipart = "^0.0.18"
PyJWT = "^2.10.1"
cohere = "^5.5.8"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
httpx = "^0.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]